Refactor app architecture to use view models
Move the ORT related business logic out of the composable scope by
introducing view models, similar to how view models are used with
JetpackCompose on Android, but without the complex lifecycle that is
required on Android.

This makes it a lot easier to control which parts of the UI need to be
updated if the configuration changes, and better decouples the UI from
the business logic.

While at it, update the view model on changes instead of only when a
screen is loaded to further simplify the code. This is possible because
it turned out that even for large ORT results the updates do not take
much time.

Also organize the files that belong together hierarchically below the
ui package instead of in separate packages for better overview.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>