Fix issue with FileDialog reopening
If the file in the `FileDialog` is selected with a double-click it can
happen that the dialog immediately opens again. The root cause seems to
be that another click event in Compose is emitted. Work around this
issue by opening the dialog only if the `OrtModel` is not currently
loading or processing a result file.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>