OrtModel: Fix reloading the same file
When reloading the same file the `_ortResult` `StateFlow` was assigned
the same value as before and as a result no new value got emitted. This
resulted in the UI getting stuck because the state was never updated
from `LOADING_RESULT`.

Fix this by resetting `_ortResult` to `null` before assigning any new
value.

Fixes #49.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>