Change app layout
Change the app layout as a preparation for introducing a navigation
backstack and adding section for the ORT tools (analyzer, scanner, ...).

This introduces a so far unused title bar, changes the background color
of the menu and adds a light gray background to the content area. The
design will further be improved in follow-up commits.

Please note that this commit temporarily breaks the dark theme. This
will be fixed in a later commit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>