refactor: Migrate to the new navigation
Migrate to the new `NavController` and do several refactorings along the
way which are required for the migration. View models are now managed by
the `NavController` and automatically closed when a screen is removed
from the backstack. This will make it easier in the future to implement
navigation for parameterized subscreens like the package details screen.

The `PackageDetailsViewModel` is already implemented as a
`MoleculeViewModel`, the other view models will be migrated later on.
This will mainly have the benefit that it makes the data flow more
obvious by consuming events in the view model and producing immutable
states as output. This is a bit similar to using Redux [1] in react
applications.

[1]: https://redux.js.org

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>