refactor: Inline filter data with filter classes
Inline the `FilterData` instances with the respective filter classes to
remove the duplication of the current filter selection. This also allows
to simplify the view model initialization a bit. For example, in case of
the `IssuesViewModel` an update of the `issues` property now directly
triggers an update of `filter` which then triggers an update of `state`.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>