fix: Prevent file dialog from reopening
Fix the issue again that the file dialog could reopen when selecting the
file with a double-click. This was originally fixed in f4c4160 by
checking if another file is already loading before showing the dialog.
The check was removed in 47d7455 and it also did not make sense anymore
to do this check in `openResultDialog` because since 871c2c2 there can
be more than one instance of `OrtModel`. Therefore reintroduce the check
in the `loadResult` function which opens the file dialog.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>