feat: Add an option to use only the resolved configuration
Add an option to use only the resolved configuration from the ORT result
file instead of the configuration from the local ORT config directory.
In the first iteration this only affects package configurations and
resolutions. Support for package curations will be added later.

Additional resolved configurations can be supported once implemented in
ORT [1].

The goal of this feature is to be able to use the ORT Workbench as a
replacement for a generated report. Currently this is not the case, as
the local configuration can be different to the one used by the ORT
tools which can lead to inconsistencies.

Note that currently not all screen update correctly when the setting is
changed, for example the package details screen needs to be reloaded to
show the changes and all filters are reset. This will be addressed in
later commits.

[1]: https://github.com/oss-review-toolkit/ort/issues/7453

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>