feat(dependencies): Preserve the dependency tree state on changes
Preserve the state of the dependency tree (e.g. current search or
expanded items) when the model changes, for example when the user
changes the setting to only use the resolved configuration.

To achieve this, add a new function to the `TreeState` that can handle
an update of the tree nodes. Also change the key used to remember the
effective license for a package to ensure it gets properly updated when
the license information changes.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@gmail.com>